backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.8.1
google-generativeai==0.3.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2